feat: move login to auth
