+++ src/controller/userController.ts
@@ -1,30 +1,3 @@
-import { Request, Response } from 'express';
-
-import { rm, sc } from '../constants';
-import { fail, success } from '../constants/response';
-import jwtHandler from '../modules/jwtHandler';
-
-//* 로그인
-const signInUser = async (req: Request, res: Response) => {
-  try {
-    const userId = 4; // 더미 유저의 id
-    const accessToken = jwtHandler.sign(userId);
-
-    const result = {
-      id: userId,
-      accessToken,
-    };
-
-    res.status(sc.OK).send(success(sc.OK, rm.SIGNIN_SUCCESS, result));
-  } catch (e) {
-    // console.log(error);
-    // ? 서버 내부에서 오류 발생
-    res
-      .status(sc.INTERNAL_SERVER_ERROR)
-      .send(fail(sc.INTERNAL_SERVER_ERROR, rm.INTERNAL_SERVER_ERROR));
-  }
-};
-
-const userController = { signInUser };
+const userController = {};
 
 export default userController;
