+++ src/router/authRouter.ts
@@ -1,5 +1,9 @@
 import { Router } from 'express';
 
+import { authController } from '../controller';
+
 const router: Router = Router();
+
+router.post('/login', authController.signInUser);
 
 export default router;
