+++ src/router/userRouter.ts
@@ -1,9 +1,5 @@
 import { Router } from 'express';
 
-import { userController } from '../controller';
-
 const router: Router = Router();
-
-router.post('/login', userController.signInUser);
 
 export default router;
